Fixed a major bug
Resolved an issue where the last few seconds of a song's beat were not being counted. This occurred because, although the song was paused and a transition was happening, the game continued to check for missed beats when it shouldn't have. Consequently, the last few beats of a song were initialized as -1 (instead of 0) before the song began.