Making the game fair
The staff is now more powerful and can gain more projectiles.

The sword is now larger to make it more viable as an option.

The spawner now spawns orcs on waves 1 and 2. It will then spawn skeletons on waves 3 and 4. Then it will spawn archers on waves 5 and 6. Then it will start spawning bosses on waves 7 and 8